Move utility into utils module.
